+++ find_mistakes/structure/check_structure.py
