+++ requirements.txt
@@ -1,0 +1,5 @@
+gensim
+tqdm
+alphabet_detector
+pymysql==0.9.2
+mysql
